feat(gui): GUI MVP foundation — shared @sparq/client TS package + Tauri 2 scaffold (sq-2e93) [OPUS-4.8]
Implements the FOUNDATION of the GUI epic per research/gui-design.md, in two parts.

(A) packages/sparq-client — a framework-agnostic shared-TS package that becomes the ONE
declaration of the WASM `Store` surface (WasmStore/WasmSolutionCursor/WasmModule), the
loaders (loadSparq/prewarmSparq with a configurable basePath), the SPARQL-JSON + SHACL
shapes, and the framework-agnostic query helpers (matchQuads/countQuads/streamQueryRows/
sparqShaclValidate/formatTerm). This kills the hand-redeclared WasmStore drift the design
flags (§0/§4): the site's site/src/lib/sparq-wasm.ts now RE-EXPORTS this package (every
existing @/lib/sparq-wasm import resolves unchanged) and keeps only the 3 dataset-coupled
helpers (loadIntoStore/storeToNQuads/datasetSize). The site consumes it via a tsconfig path
alias + a next.config webpack alias — no repo-root workspaces, no new install, no lockfile
change. Part (A) is the load-bearing, framework-agnostic deliverable: it transfers to any
GUI choice.

(B) gui/src-tauri — a Tauri 2 desktop SCAFFOLD (the design's framework recommendation):
Cargo.toml (standalone crate root, own [workspace], excluded from the main workspace gate),
tauri.conf.json, a minimal main window, and engine.rs — the DIRECT native engine link
(sparq-engine + sparq-core) whose commands mirror the wasm Store surface
(load/query/queryQuads/updateInPlace/explain/count/ask), backed by the full native store.
A path-scoped, non-merge-queue, ADVISORY .github/workflows/gui.yml lane (every action
SHA-pinned) installs the webkit2gtk system deps and builds/lints/typechecks the GUI; it is
DELIBERATELY NOT in the required ci-summary gate.

Evidence: site `npm run build` green with the shared package (all routes emitted) + site
lint + site tsc clean; @sparq/client tsc clean; the engine API calls in engine.rs
compile-checked against the real sparq-engine/sparq-core; cargo metadata resolves the full
Tauri graph (467 pkgs, tauri 2.11.3) alongside the engine path deps; gui.yml actionlint-
clean. HONEST: the Tauri app is a SCAFFOLD (config parses, metadata resolves, engine layer
unit-tested natively), NOT a working app — a full cargo build needs webview system libs and
is CI-validated only on this box. ZK/MPC framing inherited unchanged (research-grade, not
externally audited).

Follow-ups: sq-06gq (npm workspaces + generated .d.ts re-export), sq-9vw5 (basePath
env-switch for the webview), sq-2e93 per-platform CI matrix tracked in sq- follow-up.
