+++ gui/src-tauri/build.rs
@@ -1,0 +1,5 @@
+// [OPUS-4.8] sq-2e93 — Tauri build script: generates the context (parses tauri.conf.json,
+// embeds the frontend dist + capabilities). Standard Tauri 2 scaffold.
+fn main() {
+    tauri_build::build();
+}
